Dateien hochladen nach ''
This repository contains EEG raw data, behavioral data and source data for the publication

Rohe, T., Ehlis, A.-C. & Noppeney, U. The neural dynamics of hierarchical Bayesian causal inference in multisensory perception. Nature Communications, accepted (2019)

The file “Behavioral data.zip” contains the behavioral data from the study as well as Matlab code to fit the Bayesian Causal Inference model to the behavioral data. Please see the readme in the .zip file for details.

The file “Source data.zip” contains the individual data underlying the figures and the tables of the publication. Please see the readme in the .zip file for details.

The file “EEG raw data.zip” contains the individual raw EEG data of the publication. Please see the readme in the .zip file for details.